Remove /after directory
People currently cannot to use this plugin for things like syntax
highlighting and indentation along with a local ftdetect for JSX files
because the ftdetect in this plugin is nested in the `/after` directory,
causing it to run after the local one.

This commit addresses this by un-nesting the files from the `/after`
directory. Incidentally, this is the structure that other similar
plugins, such as vim-javascript [1], use.

[1]: https://github.com/pangloss/vim-javascript

With this change, I removed the manual installation instructions from
the readme since they no longer made sense and updating them to make sense
seemed like more effort than it was worth for the benefit. I assume that
people who are manually installing Vim plugins will know how to do that
in this case.

Fixes #29.