Don't let ':' terminate JSX classname
Flow uses syntax like `<T: Object>` to say "generic in type T, provided
that it's an object."

JSX technically allows ':' in a "JSXNamespacedName" (see [here][1]).
However, this is rarely ever used, and even still, a ':' can't be used
to terminate a namespaced name. This commit explicitly checks for the
':' at the end.

[1]: https://facebook.github.io/jsx/